refactor: adapt achr.py and optgp.py to use step() from core.py
